Add atomic download (#30)
* Add error handling for file download
* Make download atomic by renaming
* Add timeout without panicking
* Fix bug downloading header only
* Reduce error verbosity

### Problem
File download may fail for any number of reasons. Because the file now exists (in a possibly corrupted state), it will not be overwritten in subsequent runs.

### Solution
We solve this by downloading to a temporary file in the same directory, then renaming it to the actual file name or deleting it upon failure. Renaming should be atomic in POSIX systems for the same directory.

### Implementation
Some additional fixes were added in this rewrite
- Refactor file download to `async fn(tmp_path, real_path, ...) -> Result`
- Using the `Result`, we can rename if `Ok` or delete `tmp_path` if `Err` + `eprintln!()` the error
- Use `fork!()` macro to download files with higher concurrency
- Timeout for getting of course/folder/file metadata only, because file downloads may be huge and we can't bound download time
- Change `unwrap` panics into returned `Err`, since panicking threads are messy to handle

Fixes #29

---------

Co-authored-by: Walter Kong <walter.98.kong@gmail.com>